finished collision_event

+++ src/mc/mc_base_particle.rs
@@ -53,9 +53,9 @@ pub struct MCBaseParticle<T: Float> {
     /// Species of the particle
     pub species: Species,
     /// Current domain in the spatial grid (should be usize?)
-    pub domain: u32,
+    pub domain: usize,
     /// Current cell in the current domain (should be usize?)
-    pub cell: u32,
+    pub cell: usize,
 }
 
 impl<T: Float> MCBaseParticle<T> {
@@ -95,8 +95,8 @@ impl<T: Float> MCBaseParticle<T> {
     /// Return the current particle's location.
     pub fn get_location(&self) -> MCLocation {
         MCLocation {
-            domain: self.domain,
-            cell: self.cell,
+            domain: self.domain as u32,
+            cell: self.cell as u32,
             facet: 0,
         }
     }
