+++ src/mc/mc_particle.rs
@@ -11,7 +11,7 @@ use super::{
 };
 
 /// Structure used to represent a particle.
-#[derive(Debug, Default)]
+#[derive(Debug, Default, Clone)]
 pub struct MCParticle<T: Float> {
     /// Current position
     pub coordinate: MCVector<T>,
@@ -54,9 +54,9 @@ pub struct MCParticle<T: Float> {
     //// Current energy group the particle belong to (should be usize?)
     pub energy_group: usize,
     /// Current domain in the spatial grid (should be usize?)
-    pub domain: u32,
+    pub domain: usize,
     /// Current cell in the current domain (should be usize?)
-    pub cell: u32,
+    pub cell: usize,
     /// Facet to be crossed? (should be usize?)
     pub facet: u32,
     /// When crossing a facet, keep the surface normal dot product
@@ -104,8 +104,8 @@ impl<T: Float + FromPrimitive> MCParticle<T> {
     /// Returns the location of the particle as a [MCLocation] object.
     pub fn get_location(&self) -> MCLocation {
         MCLocation {
-            domain: self.domain,
-            cell: self.cell,
+            domain: self.domain as u32,
+            cell: self.cell as u32,
             facet: self.facet,
         }
     }
